+++ src/media/audio/decode/audio_decoder.h
@@ -26,7 +26,7 @@ class AudioDecoder {
  public:
   int Init();
 
-  int Decode(const uint8_t *data, int size,
+  int Decode(const uint8_t *data, size_t size,
              std::function<void(uint8_t *, int)> on_receive_decoded_frame);
 
   std::string GetDecoderName() { return "Opus"; }
